+++ frontend/src/App.vue
@@ -58,4 +58,3 @@ export default defineComponent({
         <KernXWebExecutor ref="kernXExecutor" />
     </div>
 </template>
-